Refactor code by moving TablePanel to tables package
This commit refactors the code by moving the TablePanel class to a new tables package to group all the classes related to tables. The class has been renamed to reflect its new package location. Additionally, the import of FilterPanel has been added to TablePanel. Finally, the references to TablePanel from ParametersPanel and ParserPanel have been updated to reflect the new package structure.
